+++ app/src/main/java/com/example/mentos/compassapp/views/CompassView.kt
@@ -9,21 +9,21 @@ import kotlinx.android.synthetic.main.view_compass.view.*
 
 class CompassView(context: Context, attrs: AttributeSet) : FrameLayout(context, attrs) {
 
-    private var azimuth: Int = 0
-    private var bearing: Int? = null
+    private var azimuth: Float = 0f
+    private var bearing: Float? = null
 
     init {
         inflate(context, R.layout.view_compass, this)
     }
 
-    fun setAzimuth(azimuth: Int) {
+    fun setAzimuth(azimuth: Float) {
         this.azimuth = azimuth
-        imageCompass.rotation = -azimuth.toFloat()
-        imageArrow.rotation = imageCompass.rotation + (bearing ?: 0)
+        imageCompass.rotation = -azimuth
+        imageArrow.rotation = imageCompass.rotation + (bearing ?: 0f)
         tvAzimuth.text = context.getString(R.string.compass_view_azimuth, azimuth)
     }
 
-    fun setBearing(bearing: Int) {
+    fun setBearing(bearing: Float) {
         this.bearing = bearing
         imageArrow.isVisible = true
         imageArrow.rotation = imageCompass.rotation + bearing
